requirements.txt
streamlit
transformers
torch
Pillow